update type file
