+++ src/order/get-order.ts
@@ -3,16 +3,13 @@ import config from "../config";
 import getAuthToken from "../auth/get-auth-token";
 
 import type { OrderResponseBody } from "@paypal/paypal-js";
-import type {
-  OrderErrorResponse,
-  GetHTTPStatusCodeSuccessResponse,
-  GetOrderResponse,
-  HttpErrorResponse,
-} from "./order";
+import type { OrderErrorResponse, HttpErrorResponse } from "./order";
 
 const {
   paypal: { apiBaseUrl },
 } = config;
+
+type GetHTTPStatusCodeSuccessResponse = 200;
 
 type GetOrderRequestHeaders = {
   Authorization: string;
@@ -24,6 +21,18 @@ type GetOrderOptions = {
   headers?: GetOrderRequestHeaders;
   query?: { fields: string };
 };
+
+type GetOrderResponse =
+  | {
+      status: "ok";
+      data: OrderResponseBody;
+      httpStatusCode: GetHTTPStatusCodeSuccessResponse;
+    }
+  | {
+      status: "error";
+      data: OrderErrorResponse;
+      httpStatusCode: Omit<number, GetHTTPStatusCodeSuccessResponse>;
+    };
 
 export default async function getOrder({
   orderID,
