+++ src/order/order.d.ts
@@ -14,10 +14,6 @@ export type OrderErrorResponse = {
 
 type HTTPStatusCodeSuccessResponse = 200 | 201 | 204 | 207;
 
-export type GetHTTPStatusCodeSuccessResponse = 200;
-
-export type PatchHTTPStatusCodeSuccessResponse = 204;
-
 // for create order and capture order
 export type CHTTPStatusCodeSuccessResponse = 200 | 201;
 
@@ -51,41 +47,6 @@ export type OrderResponse =
       httpStatusCode: Omit<number, HTTPStatusCodeSuccessResponse>;
     };
 
-export type PatchOrderResponse =
-  | {
-      status: "ok";
-      data: null;
-      httpStatusCode: PatchHTTPStatusCodeSuccessResponse;
-    }
-  | {
-      status: "error";
-      data: OrderErrorResponse;
-      httpStatusCode: Omit<number, HTTPStatusCodeSuccessResponse>;
-    };
-
-export type GetOrderResponse =
-  | {
-      status: "ok";
-      data: OrderResponseBody;
-      httpStatusCode: GetHTTPStatusCodeSuccessResponse;
-    }
-  | {
-      status: "error";
-      data: OrderErrorResponse;
-      httpStatusCode: Omit<number, HTTPStatusCodeSuccessResponse>;
-    };
-
 export type HttpErrorResponse = {
   statusCode?: number;
 } & Error;
-
-export type PatchOrderRequestBody = {
-  patch_request: PatchRequest[];
-};
-
-export type PatchRequest = {
-  op: "add" | "remove" | "replace" | "move" | "copy" | "test";
-  from?: string;
-  path: string;
-  value: number | string | boolean | null | object;
-};
